Fix date/time setting
This is actually fairly involved, because I took the opportunity to
stash all of these things into a state struct that gets threaded through
the place, simply because there are likely quite a few things that will
need to go onto there in the future.

It also helps because it both cuts down on the amount of globals
floating around, and makes the code (mostly) re-entrant.