Add string interning
Adds a new source file with string interning features.

Essentially, string interning is taking variable-length character
strings, and storing in a table that is referred to with integer keys.

This allows you to cut down on storage space for strings, as well as do
comparisons between interned strings near instantly.

Currently, this isn't used anywhere. However, it will be used for
process names, at least.