'Fix' priority queues
I say 'fix' in quotes, because we actually just cheated, and used a
seperate queue for each priority level. :-|

On the plus side, if we ever want to do nested priorities, or to do
other things with the queues, that'll be fairly easy to implement.

All you'll have to do is create a new queuetype (QT_*) and fix all of
the switches on queue types.